+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.7.0] - 2018-10-29
 ### Changed
 - `<` inside an attribute value is an error now.
 - `Token::Declaration` represents *standalone* as `bool` now.
@@ -89,7 +91,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - `log` 0.3 -> 0.4
 
-[Unreleased]: https://github.com/RazrFalcon/xmlparser/compare/v0.6.1...HEAD
+[Unreleased]: https://github.com/RazrFalcon/xmlparser/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/RazrFalcon/xmlparser/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/RazrFalcon/xmlparser/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/RazrFalcon/xmlparser/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/RazrFalcon/xmlparser/compare/v0.4.1...v0.5.0